+++ src/App.vue
@@ -13,7 +13,6 @@
       :searchTerm="currentFilterState.searchTerm"
       :loading="loading"
       :error="error"
-      :video="videoList"
       @retry="loadVideoData"
     />
